Tests/LibWeb/Text/expected/video-failed-load.txt

failed to load: "data:"
failed to load: "file:///i-do-no-exist-i-swear"
failed to load: "https://i-do-no-exist-i-swear.net.uk/"
